🔧 fix(dev): make the DB-backed tests work without knowing they need a database
Getting the Postgres suites to run locally required knowing three things that
were written down nowhere: that TEST_POSTGRES_DSN had to be set, what to set
it to, and that the image had to be the one CI uses. Missing any of them
produced ~66 BeforeEach failures that read as broken code, and the error text
pointed at `dagger call test` — a full pipeline for what should be an
edit/test loop.

Export TEST_POSTGRES_DSN from the dev shell, next to GOEXPERIMENT, pointing at
the postgres service docker-compose already defines. Entering the directory
and running `go test ./...` now works once that service is up, with nothing to
remember. flake.nix rather than a .env file because that is where this
repository already keeps shell environment, and no repo in the forest has a
.env to copy the pattern from. The variable is only an address; CI overrides
it with its own service binding, so exporting it here cannot affect the
pipeline.

Reuse the compose service instead of a purpose-built container. It already
pins the image the Dagger pipeline binds, so a second definition would have
been a third copy of a fact that has to stay identical. What matters is the
image, not the address: it carries pgvector and pg_duckdb, and a stock
postgres passes most of the suite while failing pkg/spanembed on a missing
extension — worse than no local database, because it presents an environment
gap as a defect in a package the developer did not touch. test-db-up asserts
the compose and Dagger pins still match rather than assuming it.

Fold the two byte-identical copies of the DSN helper into internal/testdb, so
there is one definition of "is the test database usable" and one message
saying what to do about it. That message now names the command to run, which
is the actual fix for the original problem: the failure explains itself
instead of requiring this knowledge up front.

Both failure modes stay failures rather than skips. A suite that skips when it
cannot find a database is indistinguishable from one that passed, and CI would
go green having exercised none of the storage layer the first time its service
binding broke.
