+++ pkg/storage/postgres/postgres_test.go
@@ -2,39 +2,16 @@ package postgres_test
 
 import (
 	"context"
-	"errors"
 	"fmt"
-	"os"
-	"time"
 
-	"github.com/jackc/pgx/v5"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 
+	"github.com/papercomputeco/tapes/internal/testdb"
 	"github.com/papercomputeco/tapes/pkg/storage/postgres"
 )
 
-func testPostgresDSN() (string, error) {
-	dsn := os.Getenv("TEST_POSTGRES_DSN")
-	if dsn == "" {
-		return "", errors.New("TEST_POSTGRES_DSN is not set; run postgres integration tests via `dagger call test` so the Dagger Postgres service is available")
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	conn, err := pgx.Connect(ctx, dsn)
-	if err != nil {
-		return "", fmt.Errorf("connect to test postgres: %w", err)
-	}
-	defer conn.Close(context.Background())
-
-	if err := conn.Ping(ctx); err != nil {
-		return "", fmt.Errorf("ping test postgres at TEST_POSTGRES_DSN: %w", err)
-	}
-
-	return dsn, nil
-}
+func testPostgresDSN() (string, error) { return testdb.DSN() }
 
 var _ = Describe("Driver", func() {
 	Describe("NewDriver", func() {
